Update mycall.c
I had a different version of this file but since it appears in the patch I thought it will be a good idea to change it so that it matches. I also added 1 to the month because the range is from 0 to 11 so it was showing 1 instead of 2 for february.